Refactor: Unsplit and decorrelate now lives in its own module.
